requirements.txt
pandas==2.3.2
pydrive2==1.21.1
google-auth==2.29.0
google-auth-httplib2==0.2.0
pyarrow==17.0.0
requests==2.32.3
supabase==2.23.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
